pyproject.toml
[project]
name = "interfaceautomation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-milvus>=0.2.1",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.6.4",
    "playwright>=1.54.0",
    "requests>=2.32.4",
]
